+++ CHANGELOG.md
@@ -11,10 +11,12 @@ All notable changes to this project will be documented in this file.
 - Support for Hubcast
 - GitLab CI for Hubcast
 - Simulation tests and generic tools for LLM MCP Tool tests
+- DOI link to README
 
 ### Changed
 - Refactored the `JobMonitorServer` and `ProfessorServer` to utilize `BaseMCPServer.run_tool()`
 - Dropped support for Python 3.10
+- Updated how plugins are discovered and registered to extend capabilities past just MCP servers
 
 ### Fixed
 -